refactor: modularize app.js into feature-specific modules
