+++ src/app.js
@@ -1,9 +1,15 @@
+import { updateForecastHeaderWithLocation, updateCurrentConditions, cToF } from './forecast.js';
+import { renderFavorites, pinCurrentLocation, editFavorite, deleteFavorite } from './favorites.js';
+import { updateRadarLayer } from './radar.js';
+import { copyForecastLimited, copyAlerts } from './clipboard.js';
+import { searchLocation } from './search.js';
+import { openSpcWindow } from './spc.js';
+
 // Global variables
 let map;
 let selectedLocation = null;
 let currentFavorites = [];
 let currentMarker = null;
-let radarLayer;
 
 window.addEventListener('DOMContentLoaded', async () => {
   // Initialize Leaflet map and expose globally
@@ -11,8 +17,8 @@ window.addEventListener('DOMContentLoaded', async () => {
   L.tileLayer('https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png', {
     attribution: '&copy; OpenStreetMap contributors'
   }).addTo(map);
-  updateRadarLayer();
-  setInterval(updateRadarLayer, 5 * 60 * 1000);
+  updateRadarLayer(map);
+  setInterval(() => updateRadarLayer(map), 5 * 60 * 1000);
   console.log("Map initialized");
 
   // DOM elements
@@ -25,10 +31,20 @@ window.addEventListener('DOMContentLoaded', async () => {
     return;
   }
 
+  window.copyForecast3 = () => copyForecastLimited(3);
+  window.copyForecast5 = () => copyForecastLimited(5);
+  window.copyForecastAll = () => copyForecastLimited(7);
+  window.copyAlerts = copyAlerts;
+  window.openSpcWindow = openSpcWindow;
+
   // Load favorites
   try {
     currentFavorites = await window.api.loadFavorites();
     renderFavorites(currentFavorites);
+    // Make favorites-related functions available to index.html
+    window.pinCurrentLocation = () => pinCurrentLocation(selectedLocation, currentFavorites, renderFavorites);
+    window.editFavorite = (index) => editFavorite(index, currentFavorites, renderFavorites);
+    window.deleteFavorite = (index) => deleteFavorite(index, currentFavorites, renderFavorites);
   } catch (err) {
     console.error('Error loading favorites:', err);
   }
@@ -130,7 +146,39 @@ window.addEventListener('DOMContentLoaded', async () => {
   document.getElementById("locationSearch").addEventListener("keydown", function (event) {
     if (event.key === "Enter") {
       event.preventDefault();
-      window.searchLocation();
+
+      window.searchLocation = () =>
+        searchLocation(
+          map,
+          { value: currentMarker },
+          (lat, lon) => { selectedLocation = { lat, lng: lon }; },
+          (el, forecast) => {
+            if (forecast?.length) {
+              const html = forecast.map(p => `
+                <div>
+                  <strong>${p.name}:</strong>
+                  <p>${p.detailedForecast}</p>
+                </div>
+              `).join('');
+              el.insertAdjacentHTML('afterend', `<h3>🌤️ 7-Day Forecast</h3>${html}`);
+            }
+          },
+          (el, alerts) => {
+            if (alerts?.length) {
+              const html = alerts.map(alert => `
+                <div>
+                  <strong>${alert.properties.headline}</strong>
+                  <p>${alert.properties.description}</p>
+                </div>
+              `).join('');
+              el.innerHTML = `<h3>Active Alerts</h3>${html}`;
+            }
+          },
+          updateForecastHeaderWithLocation,
+          updateCurrentConditions
+        );
+
+      window.searchLocation(); // Trigger search immediately
     }
   });
 
@@ -149,7 +197,7 @@ window.addEventListener('DOMContentLoaded', async () => {
       darkModeToggleButton.textContent = "🌙"; // Moon icon for light mode
     }
 
-    darkModeToggleButton.addEventListener("click", function() {
+    darkModeToggleButton.addEventListener("click", function () {
       document.body.classList.toggle("dark-mode");
 
       if (document.body.classList.contains("dark-mode")) {
@@ -162,36 +210,6 @@ window.addEventListener('DOMContentLoaded', async () => {
     });
   }
 });
-
-// Render favorites
-function renderFavorites(favorites) {
-  const favoritesDiv = document.getElementById('favorites');
-  const searchInput = document.getElementById('favoriteSearch');
-  if (!favoritesDiv) return;
-
-  let filtered = favorites;
-
-  if (searchInput && searchInput.value.trim() !== "") {
-    const query = searchInput.value.trim().toLowerCase();
-    filtered = favorites.filter(fav => fav.name.toLowerCase().includes(query));
-  }
-
-  filtered.sort((a, b) => a.name.localeCompare(b.name));
-
-  favoritesDiv.innerHTML = filtered.map((fav, i) => `
-    <div class="favorite-item">
-      <div class="favorite-content">
-        <span class="favorite-name" onclick="window.goToFavorite(${fav.lat}, ${fav.lng})">
-          ${fav.name}
-        </span>
-        <div class="favorite-actions">
-          <button class="edit-btn" onclick="window.editFavorite(${i})">✏️</button>
-          <button class="delete-btn" onclick="window.deleteFavorite(${i})">❌</button>
-        </div>
-      </div>
-    </div>
-  `).join('');
-}
 
 // Jump to favorite and fetch data
 window.goToFavorite = async function (lat, lng) {
@@ -264,212 +282,6 @@ window.goToFavorite = async function (lat, lng) {
   }
 };
 
-// Pin the current map-selected location
-window.pinCurrentLocation = async function () {
-  if (!selectedLocation) {
-    alert("Please click on the map to select a location first!");
-    return;
-  }
-
-  const { value: name } = await Swal.fire({
-    title: 'Enter a name for this location',
-    input: 'text',
-    inputPlaceholder: 'Favorite location name',
-    showCancelButton: true
-  });
-
-  if (name) {
-    const fav = { name, lat: selectedLocation.lat, lng: selectedLocation.lng };
-    currentFavorites.push(fav);
-    try {
-      await window.api.saveFavorites(currentFavorites);
-      renderFavorites(currentFavorites);
-    } catch (err) {
-      console.error('Error saving favorites:', err);
-    }
-  }
-};
-
-// Edit a favorite
-window.editFavorite = async function (index) {
-  const fav = currentFavorites[index];
-  const { value: newName } = await Swal.fire({
-    title: 'Edit Favorite',
-    input: 'text',
-    inputValue: fav.name,
-    showCancelButton: true,
-    inputPlaceholder: 'Enter new name'
-  });
-
-  if (newName && newName !== fav.name) {
-    currentFavorites[index].name = newName;
-    try {
-      await window.api.saveFavorites(currentFavorites);
-      renderFavorites(currentFavorites);
-    } catch (err) {
-      console.error('Error saving favorites:', err);
-    }
-  }
-};
-
-// Delete a favorite
-window.deleteFavorite = function (index) {
-  currentFavorites.splice(index, 1);
-  window.api.saveFavorites(currentFavorites)
-    .then(() => renderFavorites(currentFavorites))
-    .catch(err => console.error('Error saving favorites:', err));
-};
-
-// Location search
-window.searchLocation = async function () {
-  const input = document.getElementById('locationSearch');
-  const query = input.value.trim();
-  if (!query) return;
-
-  const forecastDiv = document.getElementById('forecast');
-  const alertsDiv = document.getElementById('alerts');
-
-  // Create or grab the spinner.
-  let spinner = document.getElementById('forecast-loading');
-  if (!spinner) {
-    spinner = document.createElement('div');
-    spinner.id = 'forecast-loading';
-    spinner.classList.add('spinner', 'hidden');
-    forecastDiv.prepend(spinner);
-  }
-
-  // Clear forecastDiv while preserving the spinner.
-  Array.from(forecastDiv.children).forEach(child => {
-    if (child.id !== 'forecast-loading') {
-      child.remove();
-    }
-  });
-  // Clear alerts.
-  alertsDiv.innerHTML = '';
-
-  // Show the spinner.
-  spinner.classList.remove("hidden");
-
-  const url = `https://nominatim.openstreetmap.org/search?format=json&q=${encodeURIComponent(query)}`;
-
-  try {
-    const response = await fetch(url);
-    const data = await response.json();
-
-    if (data.length > 0) {
-      const lat = parseFloat(data[0].lat);
-      const lon = parseFloat(data[0].lon);
-      selectedLocation = { lat, lng: lon };
-
-      map.setView([lat, lon], 10);
-      if (currentMarker) {
-        map.removeLayer(currentMarker);
-      }
-      currentMarker = L.marker([lat, lon]).addTo(map);
-
-      // Optionally, force a repaint with a short delay.
-      await new Promise(resolve => setTimeout(resolve, 0));
-
-      const weatherForecast = await window.api.fetchWeather(lat, lon);
-      const activeAlerts = await window.api.fetchAlerts(lat, lon);
-      await updateForecastHeaderWithLocation(lat, lon);
-      await updateCurrentConditions(lat, lon);
-
-      if (weatherForecast?.length) {
-        const forecastHtml = weatherForecast.map(period => `
-          <div>
-            <strong>${period.name}:</strong>
-            <p>${period.detailedForecast}</p>
-          </div>
-        `).join('');
-        // Append forecast content after the spinner.
-        spinner.insertAdjacentHTML('afterend', `<h3>🌤️ 7-Day Forecast</h3>${forecastHtml}`);
-      }
-
-      if (activeAlerts?.length) {
-        const alertsHtml = activeAlerts.map(alert => `
-          <div>
-            <strong>${alert.properties.headline}</strong>
-            <p>${alert.properties.description}</p>
-          </div>
-        `).join('');
-        alertsDiv.innerHTML = `<h3>Active Alerts</h3>${alertsHtml}`;
-      }
-    } else {
-      alert("Location not found.");
-    }
-  } catch (err) {
-    console.error('Error searching location:', err);
-    alert("Error searching for location.");
-  } finally {
-    // Hide the spinner when done.
-    spinner.classList.add("hidden");
-  }
-};
-
-async function updateForecastHeaderWithLocation(lat, lon) {
-  try {
-    const response = await fetch(`https://api.weather.gov/points/${lat},${lon}`);
-    const data = await response.json();
-    const city = data.properties.relativeLocation.properties.city;
-    const state = data.properties.relativeLocation.properties.state;
-    const locationName = `${city}, ${state}`;
-    document.getElementById("forecast-location").textContent = `${locationName}`;
-  } catch (err) {
-    console.warn('Could not fetch location info:', err);
-    document.getElementById("forecast-location").textContent = "";
-  }
-}
-
-async function updateCurrentConditions(lat, lon) {
-  try {
-    const pointsResponse = await fetch(`https://api.weather.gov/points/${lat},${lon}`);
-    const pointsData = await pointsResponse.json();
-    const observationsUrl = pointsData.properties.observationStations;
-
-    const stationsResponse = await fetch(observationsUrl);
-    const stationsData = await stationsResponse.json();
-    const firstStation = stationsData.features[0]?.id;
-    if (!firstStation) throw new Error("No stations found");
-
-    const conditionsResponse = await fetch(`${firstStation}/observations/latest`);
-    const conditionsData = await conditionsResponse.json();
-    const props = conditionsData.properties;
-
-    const temp = props.temperature.value;
-    let feelsLike = null;
-    if (props.heatIndex?.value !== null) {
-      feelsLike = props.heatIndex.value;
-    } else if (props.windChill?.value !== null) {
-      feelsLike = props.windChill.value;
-    }
-    const description = props.textDescription;
-    const updated = new Date(props.timestamp);
-
-    const formatTemp = t => `${Math.round(cToF(t))}°F`;
-    const formatted = [
-      description,
-      temp !== null ? formatTemp(temp) : null,
-      feelsLike !== null && feelsLike !== temp ? `(Feels like ${formatTemp(feelsLike)})` : null
-    ].filter(Boolean).join(', ');
-    console.log("Temp:", temp, "Heat Index:", props.heatIndex?.value, "Wind Chill:", props.windChill?.value);
-
-    const time = updated.toLocaleTimeString([], { hour: 'numeric', minute: '2-digit' });
-
-    document.getElementById("current-conditions").innerHTML = `
-      ${formatted}<br>
-      <span class="updated-time">Updated at ${time}</span>
-    `;
-  } catch (err) {
-    console.warn("Could not load current conditions:", err);
-    document.getElementById("current-conditions").textContent = "";
-  }
-}
-
-function cToF(celsius) {
-  return (celsius * 9) / 5 + 32;
-}
-
 window.copyForecast3 = function () {
   copyForecastLimited(3);
 };
@@ -480,168 +292,4 @@ window.copyForecast5 = function () {
 
 window.copyForecastAll = function () {
   copyForecastLimited(7);
-};
-
-function copyForecastLimited(limit) {
-  const forecastEl = document.getElementById("forecast");
-  if (!forecastEl) return;
-
-  const lines = [];
-
-  Array.from(forecastEl.children).slice(0, limit * 2 + 2).forEach(item => {
-    const boldEl = item.querySelector("b");
-    let label = "";
-    let description = "";
-
-    if (boldEl) {
-      label = boldEl.textContent.trim().replace(/:$/, "");
-
-      let contentParts = [];
-      let currentNode = boldEl.nextSibling;
-
-      while (currentNode) {
-        if (currentNode.nodeType === Node.TEXT_NODE) {
-          contentParts.push(currentNode.textContent.trim());
-        } else if (currentNode.nodeType === Node.ELEMENT_NODE) {
-          contentParts.push(currentNode.textContent.trim());
-        }
-        currentNode = currentNode.nextSibling;
-      }
-
-      description = contentParts.join(" ").replace(/\s+/g, " ").trim();
-      lines.push(`${label}: ${description}`);
-    } else {
-      const raw = item.textContent.trim().replace(/\s+/g, " ");
-      if (raw) lines.push(raw);
-    }
-  });
-
-  const finalText = lines.join("\n");
-
-  navigator.clipboard.writeText(finalText)
-    .then(() => {
-      Swal.fire({
-        icon: 'success',
-        title: 'Copied!',
-        text: `Copied ${limit}-day forecast to clipboard.`,
-        timer: 1500,
-        showConfirmButton: false
-      });
-    })
-    .catch(err => {
-      console.error("Failed to copy forecast:", err);
-      Swal.fire({
-        icon: 'error',
-        title: 'Oops!',
-        text: 'Could not copy forecast. Try again.',
-      });
-    });
-}
-
-window.copyAlerts = function () {
-  const alertsEl = document.getElementById("alerts");
-  if (!alertsEl) return;
-
-  // Grab all alert blocks (assuming each alert is in its own child div)
-  const alertItems = Array.from(alertsEl.children);
-  const lines = [];
-
-  alertItems.forEach(item => {
-    const text = item.innerText.trim().replace(/\s+/g, " ");
-    if (text) lines.push(text);
-  });
-
-  if (lines.length === 0) {
-    Swal.fire({
-      icon: 'info',
-      title: 'No Alerts',
-      text: 'There are no active alerts to copy.',
-      timer: 1500,
-      showConfirmButton: false
-    });
-    return;
-  }
-
-  const finalText = lines.join("\n\n"); // Separate alerts with double newlines
-
-  navigator.clipboard.writeText(finalText)
-    .then(() => {
-      Swal.fire({
-        icon: 'success',
-        title: 'Copied!',
-        text: 'Active alerts copied to clipboard.',
-        timer: 1500,
-        showConfirmButton: false
-      });
-    })
-    .catch(err => {
-      console.error("Failed to copy alerts:", err);
-      Swal.fire({
-        icon: 'error',
-        title: 'Oops!',
-        text: 'Could not copy alerts. Try again.',
-      });
-    });
-};
-
-async function updateRadarLayer() {
-  try {
-    const res = await fetch("https://api.rainviewer.com/public/weather-maps.json");
-    const data = await res.json();
-
-    const pastFrames = data.radar.past;
-    const availableTimestamps = pastFrames.map(frame => frame.time).reverse(); // Newest to oldest
-
-    let workingTimestamp = null;
-
-    for (const timestamp of availableTimestamps) {
-      const testUrl = `https://tilecache.rainviewer.com/v2/radar/${timestamp}/256/4/3/5/2/1_1.png`;
-
-      // Try to fetch one tile from this timestamp to confirm it's real
-      const response = await fetch(testUrl, { method: 'HEAD' });
-      if (response.ok) {
-        workingTimestamp = timestamp;
-        break;
-      }
-    }
-
-    if (!workingTimestamp) {
-      console.warn("No working radar frame found.");
-      return;
-    }
-
-    if (radarLayer) {
-      map.removeLayer(radarLayer);
-    }
-
-    radarLayer = L.tileLayer(`https://tilecache.rainviewer.com/v2/radar/${workingTimestamp}/256/{z}/{x}/{y}/2/1_1.png`, {
-      opacity: 0.6,
-      maxZoom: 12,
-      attribution: 'Radar &copy; RainViewer.com'
-    });
-
-    radarLayer.addTo(map);
-    console.log("Radar layer added with working timestamp:", workingTimestamp);
-  } catch (error) {
-    console.error("Error loading radar layer:", error);
-  }
-}
-
-window.openSpcWindow = function (day) {
-  const links = {
-    day1: 'https://www.spc.noaa.gov/products/outlook/day1otlk.html',
-    day2: 'https://www.spc.noaa.gov/products/outlook/day2otlk.html',
-    day3: 'https://www.spc.noaa.gov/products/outlook/day3otlk.html',
-  };
-
-  const width = 1280;
-  const height = 1024;
-  const left = (window.screen.width / 2) - (width / 2);
-  const top = (window.screen.height / 2) - (height / 2);
-
-  window.open(
-    links[day],
-    `${day}SPCOutlook`,
-    `width=${width},height=${height},top=${top},left=${left},resizable=yes,scrollbars=yes`
-  );
 };
